refactor trainer callbacks
