+++ trl/trainer/__init__.py
@@ -19,6 +19,7 @@ from typing import TYPE_CHECKING
 from ..import_utils import _LazyModule, is_diffusers_available, OptionalDependencyNotAvailable
 
 _import_structure = {
+    "callbacks": ["RichProgressCallback", "SyncRefModelCallback"],
     "utils": [
         "AdaptiveKLController",
         "FixedKLController",
@@ -27,7 +28,6 @@ _import_structure = {
         "RunningMoments",
         "disable_dropout_in_model",
         "peft_module_casting_to_bf16",
-        "RichProgressCallback",
     ],
     "dpo_config": ["DPOConfig", "FDivergenceConstants", "FDivergenceType"],
     "dpo_trainer": ["DPOTrainer"],
@@ -62,6 +62,7 @@ else:
 
 if TYPE_CHECKING:
     # isort: off
+    from .callbacks import RichProgressCallback, SyncRefModelCallback
     from .utils import (
         AdaptiveKLController,
         FixedKLController,
@@ -70,7 +71,6 @@ if TYPE_CHECKING:
         RunningMoments,
         disable_dropout_in_model,
         peft_module_casting_to_bf16,
-        RichProgressCallback,
     )
 
     # isort: on
